drivers: support for UBI in nandflash.
This is a basic implementation of UBI.
 * there is not check for CRC header yet,
 * there is not support for fastmpap yest.

Those features will come up soon. Currently, this is an experimental
version and the code must remain light to feet into the SRAM.

The UBI support let you boot a kernel image or a bootloader and make their
upgrade safer using spare volumes. Is spreads as well the wear on
almost all NAND flash memory.

Signed-off-by: Gaël PORTAY <gael.portay@gmail.com>