drivers: support for UBI in nandflash.
This is a basic implementation of UBI.
 * there is not check for CRC header yet,
 * there is not support for fastmap yet.

Those features will come up soon. Currently, this is an experimental
version and the code must remain light to feet into the SRAM.

The UBI support let you boot a kernel or a bootloader image and make their
upgrade safier using spare volumes.

It allows to extend the UBI partition to almost the whole NAND Flash memory
(first block excepted [reserved for this bootstrap]). Thus it spreads the wear
leveling to (almost) the whole NAND flash memory.

Signed-off-by: Gaël PORTAY <gael.portay@gmail.com>